# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.APICallLog:
    properties:
      channel:
        allOf:
        - $ref: '#/definitions/entity.Channel'
        description: 关联关系
      channel_id:
        type: integer
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      duration_ms:
        type: integer
      end_time:
        type: string
      error_message:
        type: string
      id:
        example: 1
        type: integer
      method:
        type: string
      processor:
        $ref: '#/definitions/entity.MessageProcessor'
      processor_id:
        type: integer
      request_body:
        type: string
      request_headers:
        $ref: '#/definitions/entity.JSONField'
      request_id:
        type: string
      response_body:
        type: string
      response_headers:
        $ref: '#/definitions/entity.JSONField'
      response_status:
        type: integer
      start_time:
        type: string
      success:
        type: boolean
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      url:
        type: string
    type: object
  entity.Bot:
    properties:
      avatar_url:
        type: string
      bot_name:
        type: string
      bot_type:
        type: string
      channels:
        description: 关联关系
        items:
          $ref: '#/definitions/entity.Channel'
        type: array
      config:
        $ref: '#/definitions/entity.JSONField'
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      created_by:
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      description:
        type: string
      id:
        example: 1
        type: integer
      status:
        $ref: '#/definitions/entity.Status'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  entity.Channel:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      api_call_logs:
        items:
          $ref: '#/definitions/entity.APICallLog'
        type: array
      bot:
        allOf:
        - $ref: '#/definitions/entity.Bot'
        description: 关联关系
      bot_id:
        type: integer
      channel_name:
        type: string
      config:
        $ref: '#/definitions/entity.JSONField'
      connection_logs:
        items:
          $ref: '#/definitions/entity.ConnectionLog'
        type: array
      connection_status:
        $ref: '#/definitions/entity.ConnectionStatus'
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      id:
        example: 1
        type: integer
      last_connected_at:
        type: string
      messages:
        items:
          $ref: '#/definitions/entity.Message'
        type: array
      platform_type:
        type: string
      status:
        $ref: '#/definitions/entity.Status'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      webhook_path:
        type: string
    type: object
  entity.ConnectionLog:
    properties:
      channel:
        allOf:
        - $ref: '#/definitions/entity.Channel'
        description: 关联关系
      channel_id:
        type: integer
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      details:
        $ref: '#/definitions/entity.JSONField'
      error_code:
        type: string
      error_message:
        type: string
      event_type:
        type: string
      id:
        example: 1
        type: integer
      status:
        $ref: '#/definitions/entity.Status'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  entity.ConnectionStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      ConnectionStatusConnected: 已连接
      ConnectionStatusConnecting: 连接中
      ConnectionStatusDisconnected: 断开连接
      ConnectionStatusError: 连接错误
    x-enum-varnames:
    - ConnectionStatusDisconnected
    - ConnectionStatusConnected
    - ConnectionStatusConnecting
    - ConnectionStatusError
  entity.JSONField:
    additionalProperties: true
    type: object
...
  entity.MessageProcessor:
    properties:
      api_call_logs:
        items:
          $ref: '#/definitions/entity.APICallLog'
        type: array
      config:
        $ref: '#/definitions/entity.JSONField'
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      id:
        example: 1
        type: integer
      priority:
        type: integer
      processor_name:
        type: string
      processor_type:
        type: string
      routing_rules:
        description: 关联关系
        items:
          $ref: '#/definitions/entity.MessageRoutingRule'
        type: array
      status:
        $ref: '#/definitions/entity.Status'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  entity.MessageRoutingRule:
    properties:
      bot_ids:
        $ref: '#/definitions/entity.JSONField'
      channel_ids:
        $ref: '#/definitions/entity.JSONField'
      condition_logic:
        type: string
      content_patterns:
        $ref: '#/definitions/entity.JSONField'
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      id:
        example: 1
        type: integer
      is_fallback:
        type: boolean
      message_processor:
        allOf:
        - $ref: '#/definitions/entity.MessageProcessor'
        description: 关联关系
      message_types:
        $ref: '#/definitions/entity.JSONField'
      platform_types:
        $ref: '#/definitions/entity.JSONField'
      priority:
        type: integer
      processor_id:
        type: integer
      route_type:
        $ref: '#/definitions/entity.RouteType'
      rule_description:
        type: string
      rule_name:
        type: string
      sender_patterns:
        $ref: '#/definitions/entity.JSONField'
      status:
        $ref: '#/definitions/entity.Status'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  entity.RouteType:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-comments:
      RouteTypeBroadcast: 广播路由
      RouteTypeConditional: 条件路由
      RouteTypeDirect: 直接路由
      RouteTypeForward: 转发路由
    x-enum-varnames:
    - RouteTypeDirect
    - RouteTypeForward
    - RouteTypeBroadcast
    - RouteTypeConditional
  entity.Status:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      StatusActive: 激活/启用
      StatusDeleted: 已删除
      StatusInactive: 未激活/禁用
      StatusSuspended: 暂停
    x-enum-varnames:
    - StatusInactive
    - StatusActive
    - StatusDeleted
    - StatusSuspended
...
  usecase.CreateProcessorRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      description:
        type: string
      is_enabled:
        type: boolean
      name:
        type: string
      priority:
        type: integer
      type:
        type: string
    required:
    - name
    - type
    type: object
  usecase.CreateRoutingRuleRequest:
    properties:
      conditions:
        additionalProperties: true
        type: object
      is_fallback:
        type: boolean
      message_type:
        type: string
      platform_type:
        type: string
      priority:
        type: integer
      processor_id:
        type: integer
      rule_name:
        type: string
    required:
    - processor_id
    - rule_name
    type: object
...
  usecase.ProcessorListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.MessageProcessor'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  usecase.UpdateProcessorRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      description:
        type: string
      id:
        type: integer
      is_enabled:
        type: boolean
      name:
        type: string
      priority:
        type: integer
      status:
        $ref: '#/definitions/entity.Status'
    required:
    - id
    type: object
  usecase.UpdateProcessorStatusRequest:
    properties:
      id:
        type: integer
      status:
        $ref: '#/definitions/entity.Status'
    required:
    - id
    - status
    type: object
  usecase.UpdateRoutingRuleRequest:
    properties:
      conditions:
        additionalProperties: true
        type: object
      id:
        type: integer
      is_fallback:
        type: boolean
      message_type:
        type: string
      platform_type:
        type: string
      priority:
        type: integer
      processor_id:
        type: integer
      rule_name:
        type: string
      status:
        $ref: '#/definitions/entity.Status'
    required:
    - id
    type: object
...
  v1.StandardResponse:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 统一消息机器人管理平台API
  title: OmniBotGo API
  version: "1.0"
paths:
...
  /api/v1/processors:
    get:
      consumes:
      - application/json
      description: 根据条件查询处理器列表，支持分页和过滤
      parameters:
      - description: 状态
        in: query
        name: status
        type: string
      - description: 处理器类型
        in: query
        name: type
        type: string
      - default: 1
        description: 页码，默认1
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量，默认20
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/usecase.ProcessorListResult'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取处理器列表
      tags:
      - processors
    post:
      consumes:
      - application/json
      description: 创建新的消息处理器配置
      parameters:
      - description: 创建处理器请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.CreateProcessorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.MessageProcessor'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "409":
          description: 处理器名称已存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 创建新的消息处理器
      tags:
      - processors
  /api/v1/processors/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的处理器配置
      parameters:
      - description: 处理器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 处理器不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 删除处理器
      tags:
      - processors
    get:
      consumes:
      - application/json
      description: 根据ID获取指定处理器的详细信息
      parameters:
      - description: 处理器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.MessageProcessor'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 处理器不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取处理器详情
      tags:
      - processors
    put:
      consumes:
      - application/json
      description: 更新指定处理器的配置信息
      parameters:
      - description: 处理器ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新处理器请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.UpdateProcessorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.MessageProcessor'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 处理器不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 更新处理器配置
      tags:
      - processors
  /api/v1/processors/{id}/rules:
    get:
      consumes:
      - application/json
      description: 获取指定处理器的路由规则列表
      parameters:
      - description: 处理器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.MessageRoutingRule'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 处理器不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取路由规则列表
      tags:
      - processors
    post:
      consumes:
      - application/json
      description: 为指定处理器创建路由规则
      parameters:
      - description: 处理器ID
        in: path
        name: id
        required: true
        type: string
      - description: 创建路由规则请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.CreateRoutingRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.MessageRoutingRule'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 处理器不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 创建路由规则
      tags:
      - processors
  /api/v1/processors/{id}/rules/{rule_id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的路由规则
      parameters:
      - description: 处理器ID
        in: path
        name: id
        required: true
        type: string
      - description: 规则ID
        in: path
        name: rule_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 规则不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 删除路由规则
      tags:
      - processors
    put:
      consumes:
      - application/json
      description: 更新指定的路由规则配置
      parameters:
      - description: 处理器ID
        in: path
        name: id
        required: true
        type: string
      - description: 规则ID
        in: path
        name: rule_id
        required: true
        type: string
      - description: 更新路由规则请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.UpdateRoutingRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.MessageRoutingRule'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 规则不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 更新路由规则
      tags:
      - processors
  /api/v1/processors/{id}/status:
    patch:
      consumes:
      - application/json
      description: 启用或禁用指定的处理器
      parameters:
      - description: 处理器ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新处理器状态请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.UpdateProcessorStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 处理器不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 更新处理器状态
      tags:
      - processors
...
swagger: "2.0"